refactor: split and harden native memory backend
