+++ src/infra/memory/client/native/__init__.py
@@ -1,0 +1,5 @@
+"""Native memory backend package."""
+
+from .backend import NativeMemoryBackend
+
+__all__ = ["NativeMemoryBackend"]
