+++ tests/conftest.py
@@ -1,0 +1,3 @@
+import os
+
+os.environ["DEBUG"] = "false"
